requirements.txt
alabaster==0.7.12
argh==0.26.2
asn1crypto==1.3.0
astroid==2.3.3
atomicwrites==1.3.0
attrs==19.3.0
autopep8==1.4.4
Babel==2.8.0
backcall==0.1.0
bleach==3.1.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cloudpickle==1.3.0
cryptography==2.8
cycler==0.10.0
cymem==2.0.2
cytoolz==0.9.0.1
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
dill==0.2.9
docutils==0.16
entrypoints==0.3
flake8==3.7.9
future==0.18.2
idna==2.9
imagesize==1.2.0
importlib-metadata==1.5.0
intervaltree==3.0.2
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.15.2
jeepney==0.4.3
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.2.0
jupyter-client==6.1.2
jupyter-core==4.6.3
keyring==21.1.1
kiwisolver==1.1.0
lazy-object-proxy==1.4.3
llvmlite==0.31.0
MarkupSafe==1.1.1
matplotlib==3.1.3
mccabe==0.6.1
mistune==0.8.4
mkl-service==2.3.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
numba==0.48.0
numpy==1.18.1
numpydoc==0.9.2
packaging==20.3
pandas==1.0.3
pandocfilters==1.4.2
parso==0.5.2
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
plac==0.9.6
pluggy==0.13.1
preshed==2.0.1
prompt-toolkit==3.0.4
psutil==5.7.0
ptyprocess==0.6.0
pycodestyle==2.5.0
pycparser==2.20
pydocstyle==4.0.1
pyflakes==2.1.1
Pygments==2.6.1
pylint==2.4.4
pyOpenSSL==19.1.0
pyparsing==2.4.6
pyrsistent==0.16.0
PySocks==1.7.1
python-dateutil==2.8.1
python-jsonrpc-server==0.3.4
python-language-server==0.31.10
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
pyzmq==18.1.1
QDarkStyle==2.8.1
QtAwesome==0.7.0
qtconsole==4.7.3
QtPy==1.9.0
regex==2020.4.4
requests==2.23.0
rope==0.16.0
Rtree==0.9.3
scikit-learn==0.22.1
scipy==1.4.1
seaborn==0.10.0
SecretStorage==3.1.2
sentence-splitter==1.4
six==1.14.0
snowballstemmer==2.0.0
sortedcontainers==2.1.0
spacy==2.0.16
Sphinx==2.4.4
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
spyder==4.1.2
spyder-kernels==1.9.0
testpath==0.4.4
thinc==6.12.1
toolz==0.10.0
tornado==6.0.4
tqdm==4.45.0
traitlets==4.3.3
ujson==1.35
umap-learn==0.4.1
urllib3==1.25.8
watchdog==0.10.2
wcwidth==0.1.9
webencodings==0.5.1
wrapt==1.10.11
wurlitzer==2.0.0
yapf==0.28.0
zipp==2.2.0
